+++ library/class_upfront_compat.php
@@ -49,6 +49,30 @@ class Upfront_Compat implements IUpfront_Server {
 		}
 
 		add_action('wp_ajax_upfront-notices-dismiss', array($this, 'json_dismiss_notices'));
+
+		// Hummingbird compat layer
+		add_filter('wphb_minification_display_enqueued_file', array($this, 'is_upfront_resource_skippable_with_hummingbird'), 10, 3);
+		add_filter('wphb_combine_resource', array($this, 'is_upfront_resource_skippable_with_hummingbird'), 10, 3);
+		add_filter('wphb_minify_resource', array($this, 'is_upfront_resource_skippable_with_hummingbird'), 10, 3);
+	}
+
+	/**
+	 * Hooks up to Hummingbird's native filters
+	 *
+	 * Fixes the necessary resources needed for Upfront to boot properly
+	 *
+	 * @param bool $action Context-dependent action to take (false means "skip")
+	 * @param mixed $item Content-dependent item to process - array for display, string for combine/minify actions
+	 * @param string $type Item type
+	 *
+	 * @return bool
+	 */
+	public function is_upfront_resource_skippable_with_hummingbird ($action, $item, $type) {
+		$handle = is_string($item) ? $item : (!empty($item['handle']) ? $item['handle'] : 'unknown');
+		return in_array($handle, array('upfront-main', 'upfront-element-styles', 'upfront-element-scripts'))
+			? false
+			: $action
+		;
 	}
 
 	/**
@@ -74,7 +98,11 @@ class Upfront_Compat implements IUpfront_Server {
 	private function _check_v1_backup () {
 		if (!Upfront_Permissions::current(Upfront_Permissions::BOOT)) return false; // We don't care, not editable
 		if (function_exists('upfront_exporter_is_running') && upfront_exporter_is_running()) return false; // Not in exporter
-		if ($this->_is_update_notice_dismissed()) return false; // We have notices dismissed for this version and below
+		if ($this->_is_update_notice_dismissed_for('1.0')) return false; // We have notices dismissed for v1.0 version and below
+
+		if (!class_exists('Upfront_Compat_Backup_Info')) require_once('compat/class_upfront_compat_backup_info.php');
+		$info = new Upfront_Compat_Backup_Info;
+		if (!$info->is_actionable()) return false;
 
 		// This check is potentially costly, so don't do it unless we have to
 		if (!(defined('DOING_AJAX') && DOING_AJAX)) {
@@ -102,16 +130,24 @@ class Upfront_Compat implements IUpfront_Server {
 		$updated = $cache->get('upfront-updated', 'upfront-core');
 		
 		if ($updated === false) {
-			global $wpdb;
-			$theme_key = $wpdb->esc_like(Upfront_Model::get_storage_key()) . '%';
-			$global_key = $wpdb->esc_like('upfront_') . '%';
-			$result = $wpdb->get_var($wpdb->prepare("SELECT COUNT(*) FROM {$wpdb->options} WHERE option_name LIKE %s OR option_name LIKE %s", $global_key, $theme_key));
-			$updated = !empty($result) ? 'yes' : 'no';
+			$updated_flag = get_option('upfront_is_updated_install');
+			if (empty($updated_flag)) {
+				global $wpdb;
+				$theme_key = $wpdb->esc_like('_transient_' . Upfront_Model::get_storage_key()) . '%ver1.0.0'; // Check transition caches
+				$global_key = $wpdb->esc_like('upfront_') . '%'; // Check global keys
+				$result = (int)$wpdb->get_var($wpdb->prepare("SELECT COUNT(*) FROM {$wpdb->options} WHERE option_name LIKE %s OR option_name LIKE %s", $global_key, $theme_key));
 
-			$cache->set('upfront-updated', 'upfront-core', $updated);
+				$updated = !empty($result) ? 'yes' : 'no';
+
+				if (empty($result)) {
+					update_option('upfront_is_updated_install', 'no');
+				}
+
+				$cache->set('upfront-updated', 'upfront-core', $updated);
+			}
 		}
-		
-		return 'yes' === $updated;
+
+		return !empty($updated) && 'yes' === $updated;
 	}
 
 	/**
@@ -167,11 +203,14 @@ class Upfront_Compat implements IUpfront_Server {
 			'theme_url' => admin_url('themes.php'),
 		);
 
-		if (!empty($this->_has_backup_notice)) {
+		if (!empty($this->_has_backup_notice) && Upfront_Permissions::current(Upfront_Permissions::BOOT)) {
+			if (!class_exists('Upfront_Compat_Backup_Info')) require_once('compat/class_upfront_compat_backup_info.php');
+			$info = new Upfront_Compat_Backup_Info;
 			$data['Compat']['notice'] = '' .
 				__('We’ve put a lot of time into getting the migration process right, however given the variety of layouts that can be achieved with Upfront and the amazing improvements we’ve added in version 1.0, we strongly advise that you to make a full backup of your site with <b>Snapshot</b> before proceeding to edit your site. ', 'upfront') .
 			'';
-			$data['Compat']['snapshot_url'] = esc_url('https://premium.wpmudev.org/project/snapshot/');
+			$data['Compat']['snapshot_url'] = esc_url($info->get_plugin_link());
+			$data['Compat']['snapshot_msg'] = esc_html($info->get_plugin_action());
 		}
 
 		return $data;